Fixed animations for warrior and mage
Replaced old animations with new ones which will be consistent in size with all characters. Matched animation timings with movement. Added offsets for y-axis to account for the extra space in top area of animations (excess space is for the effects and jump). Archer still to be done, DO NOT SELECT ARCHER in character select.